Cart Addition
Added the cart function to add an extra step to the transaction process. I had to adjust the enumerator and some of the public variables to make some things more visible. The cart function is not iterative because that would require completely changing our code. Instead I created the createCart(uint numBoxes) function to calculate a value for the transaction based on the number of boxes sold.